One component per part
